services/api/requirements.txt
starlette==0.41.3
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
SQLAlchemy==2.0.46
alembic==1.14.0
PyJWT==2.10.1
passlib==1.7.4
PyYAML==6.0.2
